Backtester updated with three other TA stratgeies
I've moved around a lot of the code so that it is all using functions. This was to hopefully make this easier to use for people less acquainted with python. I've also added comments to the original. Next, I implemented three other strategies the long-short crossover strategy, an exponential moving average and lastly a random coin flip. Lastly, I've included a loop that runs the coin flip 1000 times just to be sure we arent getting a random lucky or unlucky result.